+++ android/src/main/java/io/sentry/react/RNSentryPackage.java
@@ -1,30 +1,46 @@
 package io.sentry.react;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
+import androidx.annotation.Nullable;
 
-import com.facebook.react.ReactPackage;
+import java.util.HashMap;
+import java.util.Map;
+
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
-import com.facebook.react.bridge.JavaScriptModule;
+import com.facebook.react.module.model.ReactModuleInfo;
+import com.facebook.react.module.model.ReactModuleInfoProvider;
+import com.facebook.react.TurboReactPackage;
 
-public class RNSentryPackage implements ReactPackage {
-    public RNSentryPackage() {
+public class RNSentryPackage extends TurboReactPackage {
+
+  @Nullable
+  @Override
+  public NativeModule getModule(String name, ReactApplicationContext reactContext) {
+    if (name.equals(RNSentryModuleImpl.NAME)) {
+      return new RNSentryModule(reactContext);
+    } else {
+      return null;
+    }
     }
 
     @Override
-    public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-        return Arrays.<NativeModule>asList(new RNSentryModule(reactContext));
+    public ReactModuleInfoProvider getReactModuleInfoProvider() {
+      return () -> {
+        final Map<String, ReactModuleInfo> moduleInfos = new HashMap<>();
+        boolean isTurboModule = BuildConfig.IS_NEW_ARCHITECTURE_ENABLED;
+        moduleInfos.put(
+            RNSentryModuleImpl.NAME,
+            new ReactModuleInfo(
+                RNSentryModuleImpl.NAME,
+                RNSentryModuleImpl.NAME,
+                false, // canOverrideExistingModule
+                false, // needsEagerInit
+                true, // hasConstants
+                false, // isCxxModule
+                isTurboModule // isTurboModule
+        ));
+        return moduleInfos;
+      };
     }
 
-    public List<Class<? extends JavaScriptModule>> createJSModules() {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-        return Collections.emptyList();
-    }
 }
