+++ android/src/newarch/java/io/sentry/react/RNSentryModule.java
@@ -1,0 +1,92 @@
+package io.sentry.react;
+
+import com.facebook.react.bridge.ReadableArray;
+import com.facebook.react.bridge.ReadableMap;
+import com.facebook.react.bridge.Promise;
+
+public class RNSentryModule extends NativeCalculatorSpec {
+
+  private final RNSentryModuleImpl impl;
+
+  RNSentryModule(ReactApplicationContext reactContext) {
+    super(reactContext);
+    this.impl = new RNSentryModuleImpl(reactContext);
+  }
+
+  @Override
+  @NonNull
+  public String getName() {
+    return RNSentryModuleImpl.NAME;
+  }
+
+  @Override
+  public void initNativeSdk(final ReadableMap rnOptions, Promise promise) {
+    this.impl.initNativeSdk(rnOptions, promise);
+  }
+
+  @Override
+  public void crash() {
+    this.impl.crash();
+  }
+
+  @Override
+  public void fetchNativeRelease(Promise promise) {
+    this.impl.fetchNativeRelease(promise);
+  }
+
+  @Override
+  public void fetchNativeAppStart(Promise promise) {
+    this.impl.fetchNativeAppStart(promise);
+  }
+
+  @Override
+  public void fetchNativeFrames(Promise promise) {
+    this.impl.fetchNativeFrames(promise);
+  }
+
+  @Override
+  public void captureEnvelope(ReadableArray rawBytes, ReadableMap options, Promise promise) {
+    this.impl.captureEnvelope(rawBytes, options, promise);
+  }
+
+  @Override
+  public void setUser(final ReadableMap user, final ReadableMap otherUserKeys) {
+    this.impl.setUser(user, otherUserKeys);
+  }
+
+  @Override
+  public void addBreadcrumb(final ReadableMap breadcrumb) {
+    this.impl.addBreadcrumb(breadcrumb);
+  }
+
+  @Override
+  public void clearBreadcrumbs() {
+    this.impl.clearBreadcrumbs();
+  }
+
+  @Override
+  public void setExtra(String key, String extra) {
+    this.impl.setExtra(key, extra);
+  }
+
+  @Override
+  public void setContext(final String key, final ReadableMap context) {
+    this.impl.setContext(key, context);
+  }
+
+  @Override
+  public void setTag(String key, String value) {
+    this.impl.setTag(key, value);
+  }
+
+  @Override
+  public void closeNativeSdk(Promise promise) {
+    this.impl.closeNativeSdk(promise);
+  }
+
+  @Override
+  public void disableNativeFramesTracking() {
+    this.impl.disableNativeFramesTracking();
+  }
+
+}
